Employees are loading!

+++ src/services/src/Demoulas.ProfitSharing.Common/Configuration/OracleHcmConfig.cs
@@ -5,5 +5,5 @@ public sealed record OracleHcmConfig
     public string? Username { get; set; }
     public string? Password { get; set; }
     public string RestFrameworkVersion { get; set; } = "9";
-    public ushort Limit { get; set; } = byte.MaxValue;
+    public byte Limit { get; set; } = byte.MaxValue;
 }
